Rename package name of exception specific functionality
